requirements_dev.txt
Click<=8.0.4; python_version < '3.7'
Click>=8.1.3; python_version >= '3.7'
coverage==5.4
Flask==2.0.3; python_version < '3.7'
Flask==2.2.5; python_version >= '3.7'
configparser==3.5.0
flake8==5.0.4
mock==3.0.5
psutil==5.9.4
pycodestyle==2.9.0
pydocstyle==4.0.1
tox==3.27.1
pytest==7.0.1; python_version < '3.7'
pytest==7.1.3; python_version >= '3.7'
pytest-cov==2.11.1
requests==2.27.1; python_version < '3.7'
requests>=2.28.0; python_version >= '3.7'
urllib3>=1.26.12
uvicorn==0.16.0; python_version < '3.7'
uvicorn>=0.19.0; python_version >= '3.7'
wheel==0.37.1; python_version < '3.7'
wheel==0.40.0; python_version >= '3.7'
markupsafe==2.0.1; python_version < '3.7'
markupsafe==2.1.2; python_version >= '3.7'
httpx==0.22.0; python_version < '3.7'
httpx==0.23.3; python_version >= '3.7'
cffi>=1.14.6
grpcio>=1.56.2 ; python_version >= '3.7'
